Added a damage upgrade, the ability to purchase its upgrade

+++ Assets/My/Scripts/Wallet/DamageToCoinTranslator.cs
@@ -17,6 +17,6 @@ public class DamageToCoinTranslator : MonoBehaviour
 
     private void OnDamageReceived(int damageTaken)
     {
-        _wallet.AddCoins(damageTaken);
+        _wallet.AddCoins((uint)damageTaken);
     }
 }
